+++ Cmdr/BuiltInCommands/Utility/me.lua
@@ -1,0 +1,13 @@
+local Players = game:GetService("Players")
+
+return {
+	Name = "me";
+	Aliases = {};
+	Description = "Displays the current player's name.";
+	Group = "DefaultUtil";
+	Args = {};
+
+	ClientRun = function()
+		return Players.LocalPlayer.Name
+	end
+}